Split source code into multiple files
Also move source files into "src" subdirectory.
